+++ detekt-cli/src/main/kotlin/io/gitlab/arturbosch/detekt/cli/CliArgs.kt
@@ -23,6 +23,14 @@ class CliArgs : Args {
 					"These filters apply on relative paths from the project root.")
 	var filters: String? = null // Using a converter for List<PathFilter> resulted in a ClassCastException
 
+	// This is currently unused and was added for experimenting with type resolution.
+	@Suppress("detekt:UnusedPrivateMember")
+	@Parameter(names = ["--classpath", "-cp"],
+			required = false,
+			hidden = true,
+			description = "Compile Classpath of the project.")
+	private var classpath: String? = null
+
 	@Parameter(names = ["--config", "-c"],
 			description = "Path to the config file (path/to/config.yml). " +
 					"Multiple configuration files can be specified with ',' or ';' as separator.")