+++ detekt-gradle-plugin/src/main/kotlin/io/gitlab/arturbosch/detekt/Detekt.kt
@@ -2,6 +2,7 @@ package io.gitlab.arturbosch.detekt
 
 import io.gitlab.arturbosch.detekt.extensions.DetektReports
 import io.gitlab.arturbosch.detekt.invoke.BaselineArgument
+import io.gitlab.arturbosch.detekt.invoke.ClasspathArgument
 import io.gitlab.arturbosch.detekt.invoke.CliArgument
 import io.gitlab.arturbosch.detekt.invoke.ConfigArgument
 import io.gitlab.arturbosch.detekt.invoke.DebugArgument
@@ -23,6 +24,7 @@ import org.gradle.api.provider.Property
 import org.gradle.api.provider.Provider
 import org.gradle.api.reporting.ReportingExtension
 import org.gradle.api.tasks.CacheableTask
+import org.gradle.api.tasks.CompileClasspath
 import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.InputFile
 import org.gradle.api.tasks.InputFiles
@@ -48,6 +50,10 @@ open class Detekt : DefaultTask() {
 	@PathSensitive(PathSensitivity.RELATIVE)
 	@SkipWhenEmpty
 	var input: ConfigurableFileCollection = project.layout.configurableFiles()
+
+	@CompileClasspath
+	@Optional
+	var classpath: ConfigurableFileCollection = project.layout.configurableFiles()
 
 	@Input
 	@Optional
@@ -124,6 +130,7 @@ open class Detekt : DefaultTask() {
 	fun check() {
 		val arguments = mutableListOf<CliArgument>() +
 				InputArgument(input) +
+				ClasspathArgument(classpath) +
 				FiltersArgument(filters.orNull) +
 				ConfigArgument(config) +
 				PluginsArgument(plugins.orNull) +